cast float32

+++ models.py
@@ -61,7 +61,7 @@ class LogNormal(tf.keras.initializers.Initializer):
     def __call__(self, shape, dtype=None):
         connmat_generator = connmat.ConnectivityMatrixGenerator(self.units, self.p, self.mean, self.stddev)
         initial_weights_mat = connmat_generator.run_generator()
-        return tf.convert_to_tensor(initial_weights_mat)
+        return tf.convert_to_tensor(tf.cast(initial_weights_mat, tf.float32))
         #normdist = tf.random.normal(shape, mean=self.mean, stddev=self.stddev, dtype=dtype)
         #return tf.math.exp(normdist)
